requirements.txt
Flask~=1.1
pyodbc
pymysql
azure.storage.blob